- moves fixture one level up in the tree - imports fixture from index file by explicitly casting fixture type to prevent TS warning - updates typess
