+++ src/api/iiif-metadata/bouwdossier/fixture.ts
@@ -1,0 +1,4 @@
+import fixture from './bouwdossier.json'
+import { Bouwdossier } from './types'
+
+export default fixture as Bouwdossier
